Refactor [#13] timeOperator -> timeManager로 변경함에 따른 변동사항
