+++ morib/src/main/java/org/morib/server/api/homeViewApi/service/start/StartTimerServiceImpl.java
@@ -2,7 +2,7 @@ package org.morib.server.api.homeViewApi.service.start;
 
 import lombok.RequiredArgsConstructor;
 import org.morib.server.domain.task.infra.TaskGateway;
-import org.morib.server.domain.timer.TimerOperator;
+import org.morib.server.domain.timer.application.TimeManager;
 import org.morib.server.domain.timer.infra.TimerGateway;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ public class StartTimerServiceImpl implements StartTimerService {
     private final TaskGateway taskGateway;
     private final TimerGateway timerGateway;
     // private final TodoOperator todoOperator
-    private final TimerOperator timerOperator;
+    private final TimeManager timeManager;
 
     @Override
     public void execute() {
