+++ morib/src/main/java/org/morib/server/domain/timer/application/TimeManager.java
@@ -1,4 +1,4 @@
-package org.morib.server.domain.timer;
+package org.morib.server.domain.timer.application;
 
 import lombok.RequiredArgsConstructor;
 import org.morib.server.domain.timer.infra.Timer;
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class TimerOperator {
+public class TimeManager {
 
     public void aggregate() {
         // 각 task의 timer들을 종합해 오늘 나의 작업시간 계산
